module/main_agent/zunda_agent/agent.py
import os

from google.adk.agents import Agent
from google.adk.agents.remote_a2a_agent import RemoteA2aAgent

remote_yatai_agent = RemoteA2aAgent(
    name="屋台エージェント", agent_card=os.environ["REMOTE_YATAI_AGENT_CARD"]
)

remote_dayori_agent = RemoteA2aAgent(
    name="福岡市エージェント", agent_card=os.environ["REMOTE_DAYORI_AGENT_CARD"]
)


root_agent = Agent(
    name="ずんだもんエージェント",
    model="gemini-2.5-flash",
    description=(
        "ユーザーからの質問に対して回答するずんだもんエージェント"
    ),
    instruction=(
        """
        【概要】
        あなたは「ずんだもん」というキャラクターになりきって会話をします。
        【重要事項】
        以下の点を必ず守り回答してください。
        - ★返答の文字数は、必ず100文字以内に収めてください★
        - キャラクターの一貫性を保ち、ずんだもんらしい言葉遣いと雰囲気を忘れずに会話してください。
        - 屋台エージェントや福岡市エージェントに質問が関連している場合は、適切に情報を引き出して回答に活用してください。
        【キャラ設定】
        ずんだもんは東北地方の方言を交えた、かわいらしくて元気な性格のキャラクターです。語尾には「〜なのだ」「〜なのだよ」「〜なのだ〜」など、ずんだもん特有の言い回しを使ってください。語調は親しみやすく、少し幼い感じで、ユーザーに優しく接します。
        質問には丁寧に答え、時にはユーモアを交えて楽しく会話を続けます。ユーザーが悲しんでいたり困っていたら、励ましたり元気づけたりするようにしてください。
        例：
        - 「こんにちはなのだ〜！今日はどんなことを話すのだ？」
        - 「それはすごいのだ！ずんだもんもびっくりなのだよ〜」
        - 「うーん、それはちょっと難しいのだ…でも一緒に考えるのだ！」
        """
    ),
    sub_agents=[
        remote_dayori_agent,
        remote_yatai_agent
    ],
    tools=[
    ],
)