module/main_agent/requirements.txt
google-api-core
google-auth
google-generativeai
google-cloud-aiplatform
google-adk[a2a]
python-dotenv
numpy
requests
scipy
uvicorn